KIALI-2279 Use the same terminology ("Fetch") and location (right side) for duration/refresh - Use "Fetch" instead of "Fetching" or "Displaying" because it is:   - consistent   - clear   - short (saves width/prevents wrapping)   - a command rather than an action - also, move RefreshContainer into Refresh component
