+++ src/components/GraphFilter/GraphRefresh.tsx
@@ -79,7 +79,7 @@ export class GraphRefresh extends React.PureComponent<GraphRefreshProps> {
 
     return (
       <>
-        <label className={GraphRefresh.durationLabelStyle}>Fetching</label>
+        <label className={GraphRefresh.durationLabelStyle}>Fetch</label>
         <ToolbarDropdown
           id={'graph_filter_duration'}
           disabled={this.props.disabled}
